chore: move ui feature function dir
